refactor: from lib to src :ocean:
